# src/GovUK.Dfe.FlexForms.Api.Client/Generated/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/Templates/{templateId}/schema": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Returns the latest template schema for the specified template name if the user has access.",
        "operationId": "Templates_GetLatestTemplateSchema",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The latest template schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSchemaDto"
                }
              }
            }
          },
          "400": {
            "description": "Request was invalid or template not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid user token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Templates/{templateId}/custom-statuses": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Get custom application statuses for a template.",
        "operationId": "Templates_GetCustomApplicationStatuses",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Custom statuses returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomApplicationStatusDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Templates"
        ],
        "summary": "Create or update a custom application status for a template.\nIf a CustomApplicationStatus for the TemplateId and ApplicationStatus exists, update its label; otherwise create one.\nReturns the created or updated CustomApplicationStatus.",
        "operationId": "Templates_CreateCustomApplicationStatus",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomApplicationStatusRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "Custom status created/updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomApplicationStatusDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/Templates/{templateId}/versions": {
      "post": {
        "tags": [
          "Templates"
        ],
        "summary": "Creates a new schema version for the specified template.",
        "operationId": "Templates_CreateTemplateVersion",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateVersionRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "Template version created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSchemaDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid user token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "InProgress",
          "Submitted",
          "Deleted",
          "Created"
        ],
        "x-enum-descriptions": [
          "In progress",
          "Submitted",
          "Deleted",
          "Created"
        ],
        "enum": [
          "InProgress",
          "Submitted",
          "Deleted",
          "Created"
        ]
      },
...
      "TemplateSchemaDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "templateId": {
            "type": "string",
            "format": "guid"
          },
          "templateVersionId": {
            "type": "string",
            "format": "guid"
          },
          "versionNumber": {
            "type": "string"
          },
          "jsonSchema": {
            "type": "string"
          }
        }
      },
      "ExceptionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "errorId": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "exceptionType": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "context": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "CustomApplicationStatusDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "customApplicationStatusId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "templateId": {
            "type": "string",
            "format": "guid"
          },
          "applicationStatus": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "CustomApplicationStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "applicationStatus",
          "label"
        ],
        "properties": {
          "applicationStatus": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "label": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "CreateTemplateVersionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "versionNumber": {
            "type": "string"
          },
          "jsonSchema": {
            "type": "string"
          }
        }
      },
...
    }
  }
}